backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
langchain==0.2.14
langgraph==0.2.14
langchain-openai==0.1.21
textblob==0.18.0.post0
pandas==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.1
pytrends==4.9.2